WIP Change the first subuid/gid mapped to root to be the user's id
When a container has the root id mapped to the first subuid in a
new user namespace it is not easy to bind mount directories. But
this is very useful for development purposes. However, if we remap
the root id to the actual user id/group then we can simply change
the owner of the directory we want to bind mount to the remap user.
Alternatively, we can change the remap user to the development
account.